Some types moved to wallet-types package, unused imports removed
