+++ common/ergvein-common.cabal
@@ -21,4 +21,5 @@ library
     , lens
     , template-haskell
     , text
+    , base64-bytestring >= 1 && < 1.0.0.3
   default-language:    Haskell2010
