+++ common/ghcid.sh
