+++ crypto/ergvein-crypto.cabal
@@ -19,11 +19,6 @@ maintainer:
                      - Aminion <bonumsignum@gmail.com>
                      - Vladimir Krutkin <krutkinvs@gmail.com>
 
-flag testnet
-  description: Build the testnet version
-  default: True
-  manual: True
-
 library
   hs-source-dirs:      src
   exposed-modules:
@@ -48,10 +43,7 @@ library
     , memory                    >= 0.14.16  && < 0.15.1
     , text                      >= 1.2      && < 1.3
     , vector                    >= 0.10     && < 0.12.1
-
     , ergvein-wallet-types      >= 0.1
-  if flag(testnet)
-    cpp-options: -DTESTNET
   default-language:    Haskell2010
   default-extensions:
     GADTs
