+++ wallet/ergvein-wallet.cabal
@@ -62,7 +62,7 @@ library
     Ergvein.Wallet.Settings
     Ergvein.Wallet.Storage
     Ergvein.Wallet.Storage.AuthInfo
-    Ergvein.Wallet.Storage.Data
+    Ergvein.Wallet.Storage.Constants
     Ergvein.Wallet.Storage.Util
     Ergvein.Wallet.Style
     Ergvein.Wallet.Util
