+++ wallet/src/Ergvein/Wallet/Client.hs
@@ -10,9 +10,7 @@ module Ergvein.Wallet.Client
   ) where
 
 import Control.Concurrent
-import Control.Concurrent.Lifted hiding (threadDelay)
 import Control.Concurrent.STM
-import Control.Concurrent.STM.TVar
 import Control.Monad.IO.Class
 import Control.Monad.Random
 import Control.Monad.Reader
@@ -27,9 +25,7 @@ import qualified Data.List as L
 import qualified Data.Set as S
 
 import Ergvein.Index.Client
-import Ergvein.Text
 import Ergvein.Wallet.Alert
-import Ergvein.Wallet.Language
 import Ergvein.Wallet.Monad.Front
 import Ergvein.Wallet.Localization.Client
