+++ wallet/src/Ergvein/Wallet/Main.hs
@@ -4,20 +4,23 @@ module Ergvein.Wallet.Main(
   ) where
 
 import Data.ByteString (ByteString)
+import Ergvein.Types.Currency
+import Ergvein.Types.Keys
 import Ergvein.Wallet.Alert.Handler
 import Ergvein.Wallet.Elements
 import Ergvein.Wallet.Loading
 import Ergvein.Wallet.Log.Writer
 import Ergvein.Wallet.Monad
+import Ergvein.Wallet.Native
 import Ergvein.Wallet.Page.Balances
 import Ergvein.Wallet.Page.Initial
 import Ergvein.Wallet.Page.Seed
 import Ergvein.Wallet.Password
 import Ergvein.Wallet.Run
 import Ergvein.Wallet.Run.Callbacks
-import Reflex.Dom.Main (mainWidgetWithCss)
+import Ergvein.Wallet.Storage.Constants
 
-import Ergvein.Wallet.Native
+import Reflex.Dom.Main (mainWidgetWithCss)
 
 frontend :: MonadFrontBase t m => m ()
 frontend = do
@@ -27,4 +30,29 @@ frontend = do
   askPasswordModal
   logWriter =<< fmap fst getLogsTrigger
   logWrite "Entering initial page"
-  void $ retractStack initialPage `liftAuth` retractStack balancesPage
+  void $ retractStack initialPage `liftAuth` (scanKeys >> retractStack balancesPage)
+
+scanKeys :: MonadFront t m => m ()
+scanKeys = mdo
+  -- gapD <- holdDyn 0 gapE
+  -- buildE <- getPostBuild
+  -- let nextE = leftmost [newE, buildE]
+  -- nextKeyE <- performEvent $ getNextKey <$> nextE
+  -- mresE <- getTxHashHistory $ TxHashHistoryRequest BTC <$> newKeyE
+  -- resE <- handleDangerMsg mresE
+  -- validE <- validateHistory resE
+  -- storedE :: Event t Int <- storeNewTransactions validE
+  -- let newE = flip push storedE $ \i -> do
+  --       gap <- sample . current $ gapD
+  --       pure $ if i == 0 && gap > gapLimit then Nothing else Just ()
+  --     gapE = flip push storedE $ \i -> do
+  --       gap <- sample . current $ gapD
+  --       pure $ if i == 0 then Just $ gap + 1 else Nothing
+  pure ()
+
+-- getNextKey :: MonadStorage t m => Currency -> KeyPurpose -> m EgvXPubKey
+-- getNextKey = _
+
+-- filterAddress :: Event t Address -> m (Event t [BlockHeight])
+
+-- getBlock :: Event t BlockHeight -> m (Event t Block)
