+++ wallet/src/Ergvein/Wallet/Monad/Storage.hs
@@ -5,11 +5,10 @@ module Ergvein.Wallet.Monad.Storage
 
 import Data.Text (Text)
 import Ergvein.Crypto
-import Control.Monad.IO.Class
+import Ergvein.Types.Storage
+import Ergvein.Types.Currency
 import Ergvein.Wallet.Monad.Base
 import Ergvein.Wallet.Native
-import Ergvein.Wallet.Storage.Data
-import Network.Haskoin.Address
 import Reflex
 
 class (MonadBaseConstr t m, HasStoreDir m) => MonadStorage t m | m -> t where
@@ -17,3 +16,4 @@ class (MonadBaseConstr t m, HasStoreDir m) => MonadStorage t m | m -> t where
   getEncryptedPrivateStorage :: m EncryptedPrivateStorage
   getWalletName              :: m Text
   storeWallet                :: Event t () -> m ()
+  -- getPubKeys                 :: m [EgvXPubKey]
